Unify how module specifiers are visited by different declaration transformers
