+++ packages/cli/src/rollup-plugins/typescript-declarations/get-used-declaration-with-replacing.ts
@@ -7,7 +7,6 @@ import {
   TS,
 } from "./common";
 import { Program, ResolvedModuleFull } from "typescript";
-import { getModuleSpecifier } from "./get-module-specifier";
 
 function replaceExt(filename: string) {
   return filename.replace(/\.([cm]?ts|tsx)$/, (match) => {
@@ -35,40 +34,18 @@ export async function getUsedDeclarationsWithPackageJsonImportsReplaced(
   );
   const emitted: EmittedDeclarationOutput[] = [];
 
-  function replaceDescendentNode<Parent extends import("typescript").Node>(
-    root: Parent,
-    oldNode: import("typescript").Node,
-    newNode: import("typescript").Node,
-    context: import("typescript").TransformationContext
-  ): Parent {
-    const visitor = (
-      node: import("typescript").Node
-    ): import("typescript").Node => {
-      if (node === oldNode) return newNode;
-      return typescript.visitEachChild(node, visitor, context);
-    };
-    return typescript.visitEachChild(root, visitor, context);
-  }
   for (const filename of depQueue) {
     const importReplacements = new Map<string, string | false>();
-    const handleImport = <Parent extends import("typescript").Node>(
-      parent: Parent,
-      moduleSpecifier: import("typescript").StringLiteral,
-      context: import("typescript").TransformationContext
-    ): Parent => {
+    const handleImport = (
+      moduleSpecifier: import("typescript").StringLiteral
+    ) => {
       const imported = moduleSpecifier.text;
       if (importReplacements.has(imported)) {
         const replacedImportPath = importReplacements.get(imported)!;
-        if (replacedImportPath === false) return parent;
-        const newStringLiteral = typescript.factory.createStringLiteral(
-          replacedImportPath
-        );
-        return replaceDescendentNode(
-          parent,
-          moduleSpecifier,
-          newStringLiteral,
-          context
-        );
+        if (replacedImportPath === false) {
+          return;
+        }
+        return typescript.factory.createStringLiteral(replacedImportPath);
       }
       const resolvedModule = resolveModuleName(imported, filename);
       if (
@@ -77,12 +54,12 @@ export async function getUsedDeclarationsWithPackageJsonImportsReplaced(
         resolvedModule.resolvedFileName.startsWith(normalizedPkgDirNodeModules)
       ) {
         importReplacements.set(imported, false);
-        return parent;
+        return;
       }
       depQueue.add(resolvedModule.resolvedFileName);
       if (imported[0] !== "#") {
         importReplacements.set(imported, false);
-        return parent;
+        return;
       }
       let forImport = replaceExt(
         normalizePath(
@@ -93,15 +70,7 @@ export async function getUsedDeclarationsWithPackageJsonImportsReplaced(
         forImport = `./${forImport}`;
       }
       importReplacements.set(imported, forImport);
-      const newStringLiteral = typescript.factory.createStringLiteral(
-        forImport
-      );
-      return replaceDescendentNode(
-        parent,
-        moduleSpecifier,
-        newStringLiteral,
-        context
-      );
+      return typescript.factory.createStringLiteral(forImport);
     };
     // this is mostly sync except for one bit so running this concurrently wouldn't really help
     const output = await getDeclarationsForFile(
@@ -111,28 +80,7 @@ export async function getUsedDeclarationsWithPackageJsonImportsReplaced(
       normalizedPkgDir,
       projectDir,
       diagnosticsHost,
-      {
-        afterDeclarations: [
-          (context) => (
-            node
-          ): import("typescript").Bundle | import("typescript").SourceFile => {
-            // typescript has a exportedModulesFromDeclarationEmit property
-            // on these source files, it's marked @internal though so i'm not using it
-            // might want to detect at runtime if it exists and use it in that case otherwise defer to this
-            // i'm not terribly worried though
-            // you should not have massive declarations files in the way that having massive source
-            // files is actually reasonable
-            const visitor = <TNode extends import("typescript").Node>(
-              node: TNode
-            ): TNode => {
-              const literal = getModuleSpecifier(node, typescript);
-              if (literal) return handleImport(node, literal, context);
-              return typescript.visitEachChild(node, visitor, context);
-            };
-            return typescript.visitEachChild(node, visitor, context);
-          },
-        ],
-      }
+      handleImport
     );
     emitted.push(output);
   }
