+++ packages/cli/src/rollup-plugins/typescript-declarations/get-used-declarations.ts
@@ -5,7 +5,6 @@ import {
   TS,
 } from "./common";
 import { Program, ResolvedModuleFull } from "typescript";
-import { getModuleSpecifier } from "./get-module-specifier";
 
 export async function getUsedDeclarations(
   typescript: TS,
@@ -33,29 +32,9 @@ export async function getUsedDeclarations(
       normalizedPkgDir,
       projectDir,
       diagnosticsHost,
-      {
-        afterDeclarations: [
-          () => (node) => {
-            // typescript has a exportedModulesFromDeclarationEmit property
-            // on these source files, it's marked @internal though so i'm not using it
-            // might want to detect at runtime if it exists and use it in that case otherwise defer to this
-            // i'm not terribly worried though
-            // you should not have massive declarations files in the way that having massive source
-            // files is actually reasonable
-            if (typescript.isSourceFile(node)) {
-              function handler(node: import("typescript").Node): void {
-                const literal = getModuleSpecifier(node, typescript);
-                if (literal) {
-                  imports.add(literal.text);
-                }
-                typescript.forEachChild(node, handler);
-              }
-              handler(node);
-              return node;
-            }
-            return node;
-          },
-        ],
+      (moduleSpecifier) => {
+        imports.add(moduleSpecifier.text);
+        return undefined;
       }
     );
     emitted.push(output);
